requirements.txt
tqdm==4.66.3
numpy==1.26.3
pandas==2.2.0
openai==1.61.1
rdkit==2023.3.1
litellm==1.60.8
anthropic==0.23.1
matplotlib==3.9.4
python-dotenv==1.0.1